# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation ",
        "description": "API Documentation",
        "contact": {
            "email": "srundavith@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/movies/{id}": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Detail",
                "description": "-",
                "operationId": "Movies/GetById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "return model admin"
                    }
                }
            },
            "put": {
                "tags": [
                    "Movies"
                ],
                "summary": "Update movies",
                "description": "-",
                "operationId": "movies/update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "form admin",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "overview",
                                    "run_time",
                                    "release_date",
                                    "total_likes",
                                    "total_ratings",
                                    "average_rating",
                                    "poster_image",
                                    "cover_image",
                                    "trailer_url",
                                    "last_upload_date",
                                    "subscription_only",
                                    "expired_subscription_only"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "overview": {
                                        "type": "string"
                                    },
                                    "run_time": {
                                        "type": "integer"
                                    },
                                    "release_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "total_likes": {
                                        "type": "integer"
                                    },
                                    "total_ratings": {
                                        "type": "integer"
                                    },
                                    "average_rating": {
                                        "type": "string"
                                    },
                                    "poster_image": {
                                        "type": "string"
                                    },
                                    "cover_image": {
                                        "type": "string"
                                    },
                                    "trailer_url": {
                                        "type": "string"
                                    },
                                    "last_upload_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "subscription_only": {
                                        "type": "string"
                                    },
                                    "expired_subscription_only": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "tags": [
                    "Movies"
                ],
                "summary": "Delete movies",
                "description": "-",
                "operationId": "movies/delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
...
        "/api/rated_movies": {
            "get": {
                "tags": [
                    "Rated_Movies"
                ],
                "summary": "Get List rated-movies Data",
                "description": "enter your rated-movies here",
                "operationId": "rated_movies",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/rated-movies": {
            "post": {
                "tags": [
                    "Rated_Movies"
                ],
                "summary": "rated-movies",
                "description": "rated-movies",
                "operationId": "Rated-Movies",
                "requestBody": {
                    "description": "form rated-movies",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "movie_id",
                                    "rated_value"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "string"
                                    },
                                    "movie_id": {
                                        "type": "string"
                                    },
                                    "rated_value": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/rated-movies/{id}": {
            "get": {
                "tags": [
                    "Rated_Movies"
                ],
                "summary": "Detail",
                "description": "-",
                "operationId": "rated-movies/GetById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "return model admin"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Rated_Movies"
                ],
                "summary": "Delete rated-movies",
                "description": "-",
                "operationId": "rated-movies/delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}